Create 14142-BridgeMST.cpp
This problem is complex. Firstly you need to find the bridges and assign right cost for each edges. Then you need to do a Minimal Spanning Tree to find the least cost. For the bridge-finding I used the Korosaju's algorithm which requires DFS the graph twice. The MST is simply Kruskal.
Also note that for online submission I used long long integer, which might be space-consuming. You might use int for some of them as well.